Made AggregateStream return nullable entity in case stream didn't have any records
