+++ EventStoreInOneHour.Tests/Exercise03CreateAppendEventFunction.cs
@@ -1,4 +1,5 @@
 using Dapper;
+using EventStoreInOneHour.Tests.BankAccounts;
 using EventStoreInOneHour.Tests.Tools;
 using FluentAssertions;
 using Npgsql;
