+++ EventStoreInOneHour.Tests/Exercise09Projections.cs
@@ -1,5 +1,6 @@
 using Dapper;
 using Dapper.Contrib.Extensions;
+using EventStoreInOneHour.Tests.BankAccounts;
 using EventStoreInOneHour.Tests.Tools;
 using FluentAssertions;
 using Npgsql;
